pyproject.toml
[project]
name = "ambuda"
version = "0.1.0"
description = "An open Sanskrit library"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "babel>=2.12.1",
    "boto3>=1.41.2",
    "celery==5.5.3",
    "defusedxml>=0.7.1",
    "dharmamitra-sanskrit-grammar>=0.1.7",
    "email-validator==1.2.1",
    "flask==3.0.0",
    "flask-babel==4.0.0",
    "flask-caching>=2.0.0",
    "flask-login==0.6.3",
    "flask-mail==0.9.1",
    "flask-wtf==1.2.1",
    "google-cloud-vision>=3.11.0",
    "google-genai==1.56.0",
    "gunicorn==20.1.0",
    "indic-transliteration==2.3.31",
    "linkify-it-py==2.0.0",
    "markdown-it-py==2.1.0",
    "markupsafe==2.1.1",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymupdf==1.26.7",
    "python-dateutil==2.9.0",
    "python-dotenv==0.20.0",
    "python-slugify==6.1.2",
    "redis==4.3.4",
    "regex==2022.6.2",
    "requests>=2.27.1",
    "sentry-sdk==2.42.0",
    "sqlalchemy==2.0.44",
    "vidyut>=0.4.0",
    "werkzeug==3.0.1",
    "wtforms==3.0.1",
    "wtforms-sqlalchemy==0.3",
]

[dependency-groups]
dev = [
    "alembic==1.17.0",
    "click==8.1.3",
    "conllu==4.5.2",
    "fabric==3.2.2",
    "moto[s3]>=5.1.17",
    "mypy-extensions==0.4.3",
    "pytest==8.4.2",
    "pytest-cov==7.0.0",
    "ruff>=0.14.1",
    "sphinx==8.2.3",
    "sphinx-rtd-theme>=1.0.0",
    "ty>=0.0.2",
]

[tool.coverage.run]
omit = [
    "ambuda/scripts/*",
    "ambuda/seed/**",
]

[tool.ruff]
# https://beta.ruff.rs/docs/rules/
lint.select = [
    "B",     # flake8-bugbear
    "F",     # pyflakes
    "N818",  # error-suffix-on-exception-name
    "UP",    # pyupgrade
    "W",     # pycodestyle
    "I001",  # isort
]
lint.ignore = [
    "B007",  # unused-loop-control-variable
    "B905"   # zip-without-explicit-strict
]
